requirement.txt
asgiref==3.8.1
attrs==24.2.0
colorama==0.4.6
Django==5.1.1
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
